services/server/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
decorator==5.1.1
Flask==2.2.5
Flask-Cors==3.0.10
frozenlist==1.3.3
gcloud==0.18.3
google-api-core==2.11.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.2
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.21.6
oauth2client==4.1.3
oauthlib==3.2.2
pandas==1.3.5
protobuf==4.23.0rc2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.29.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
typing-extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
yarl==1.9.2
zipp==3.15.0
